add policies for CompanyUser crud, add tests for it

+++ app/Http/Controllers/CompanyUserController.php
@@ -15,6 +15,8 @@ class CompanyUserController extends Controller
 {
     public function index(Company $company): View
     {
+        $this->authorize('viewAny', $company);
+
         $users = $company->users()->where('role_id', Role::COMPANY_OWNER->value)->get();
 
         return view('companies.users.index', compact(['users', 'company']));
@@ -23,11 +25,15 @@ class CompanyUserController extends Controller
 
     public function create(Company $company): View
     {
+        $this->authorize('create', $company);
+
         return view('companies.users.create', compact('company'));
     }
 
     public function store(UserStoreRequest $request, Company $company): RedirectResponse
     {
+        $this->authorize('create', $company);
+
         $validatedData = $request->validated();
         $newUser = $company->users()->create([
             'name' => $validatedData['name'],
@@ -39,18 +45,17 @@ class CompanyUserController extends Controller
         return to_route('companies.users.index', compact('company'));
     }
 
-    public function show(string $id)
-    {
-        //
-    }
-
     public function edit(Company $company, User $user): View
     {
+        $this->authorize('update', $company);
+
         return view('companies.users.edit', compact('company', 'user'));
     }
 
     public function update(UserUpdateRequest $request, Company $company, User $user): RedirectResponse
     {
+        $this->authorize('update', $company);
+
         $user->update($request->validated());
 
         return to_route('companies.users.index', compact('company'));
@@ -58,8 +63,14 @@ class CompanyUserController extends Controller
 
     public function destroy(Company $company, User $user): RedirectResponse
     {
+        $this->authorize('delete', $company);
         $user->delete();
 
         return to_route('companies.users.index', compact('company'));
     }
+
+//    public function show(string $id)
+//    {
+//        //
+//    }
 }
